pyproject.toml
[project]
name = "sovereignshield-mobile"
version = "0.1.0"
description = "SovereignShield Mobile — OPA + Claude Agents (Mobile-First)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "shiny>=0.7.1",
    "anthropic>=0.39.0",
    "supabase>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pip-audit>=2.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint.per-file-ignores]
"Artifacts/project/sovereignshield_mobile/app.py" = ["E402"]

[tool.mypy]
python_version = "3.11"
strict = true

[[tool.mypy.overrides]]
module = "supabase.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "chromadb.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sentence_transformers.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["Artifacts/project/sovereignshield_mobile/tests"]
python_files = ["test_*.py"]
markers = ["unit: mark test as unit test"]
